Add files via upload
Screenshot 2:
This screenshot shows how I played the first two clips of a stream running and birds chirping at the end. I did this to show how the forest revives itself after a disaster such as a forest fire. This also allows the soundscape to be played in a seamless loop if desired.